useMetricsData renamed to useTimeseries
