+++ src/ducks/Studio/timeseries/anomalies.js
@@ -78,3 +78,13 @@ ANOMALIES.forEach(anomaly => {
     formatter: v => v
   }
 })
+
+export function buildAnomalies (metrics) {
+  return metrics
+    .filter(({ key, anomalyKey }) => anomalyKey || ANOMALIES.includes(key))
+    .map(({ key, anomalyKey }) => ({
+      key: anomalyKey || key + '_anomaly',
+      queryKey: anomalyKey ? 'anomalies' : 'anomaly',
+      metricAnomaly: key
+    }))
+}
